fixup! fixup! JAMES-2522 Moving and renaming classes
